Heavy work on UnresolvedElementsSubProcessor and linked classes
