+++ org.eclipse.jdt.ui/ui/org/eclipse/jdt/internal/ui/text/correction/ReorgCorrectionsSubProcessor.java
@@ -537,7 +537,7 @@ public class ReorgCorrectionsSubProcessor extends ReorgCorrectionsBaseSubProcess
 		return false;
 	}
 
-	private ReorgCorrectionsSubProcessor() {
+	ReorgCorrectionsSubProcessor() {
 	}
 
 	@Override
