+++ org.eclipse.jdt.ui/ui/org/eclipse/jdt/internal/ui/text/correction/proposals/RenameNodeCorrectionProposal.java
@@ -21,6 +21,10 @@ import org.eclipse.jdt.ui.text.java.correction.CUCorrectionProposal;
 import org.eclipse.jdt.internal.ui.JavaPluginImages;
 
 public class RenameNodeCorrectionProposal extends CUCorrectionProposal {
+	public RenameNodeCorrectionProposal(RenameNodeCorrectionProposalCore core) {
+		super(core.getName(), core.getCompilationUnit(), core.getRelevance(), JavaPluginImages.get(JavaPluginImages.IMG_CORRECTION_CHANGE), core);
+	}
+
 	public RenameNodeCorrectionProposal(String name, ICompilationUnit cu, int offset, int length, String newName, int relevance) {
 		super(name, cu, relevance, JavaPluginImages.get(JavaPluginImages.IMG_CORRECTION_CHANGE), new RenameNodeCorrectionProposalCore(name, cu, offset, length, newName, relevance));
 	}
