+++ org.eclipse.jdt.ui/ui/org/eclipse/jdt/ui/text/java/correction/ASTRewriteCorrectionProposal.java
@@ -34,6 +34,18 @@ import org.eclipse.jdt.internal.ui.JavaPluginImages;
 public class ASTRewriteCorrectionProposal extends CUCorrectionProposal {
 	/**
 	 * Constructs an AST rewrite correction proposal.
+	 * @param delegate The delegate instance
+	 * @param image the image that is displayed for this proposal or <code>null</code> if no image
+	 *            is desired
+	 * @since 3.31
+	 */
+	public ASTRewriteCorrectionProposal(ASTRewriteCorrectionProposalCore delegate, Image image) {
+		super(delegate.getName(), delegate.getCompilationUnit(), delegate.getRelevance(), image, delegate);
+	}
+
+
+	/**
+	 * Constructs an AST rewrite correction proposal.
 	 * @param name the display name of the proposal
 	 * @param cu the compilation unit that is modified
 	 * @param rewrite the AST rewrite that is invoked when the proposal is applied or
